odometerInterface gets double instead of int

+++ src/main/java/bananacore/epic/OurParser.java
@@ -34,7 +34,7 @@ public class OurParser {
             carController.updateFuelLevel(value, timestamp);
         }
     }
-    private void updateOdometerObservers(int value, Timestamp timestamp){
+    private void updateOdometerObservers(double value, Timestamp timestamp){
         for (OdometerInterface carController : odometerObservers) {
             carController.updateOdometer(value, timestamp);
         }
@@ -141,7 +141,7 @@ public class OurParser {
         }else if (name.equals("transmission_gear_position")){
             updateGearObservers(numericToInt((String) jsonObject.get("value")),timestamp);
         }else if (name.equals("odometer")){
-            updateOdometerObservers((int) Double.parseDouble(jsonObject.get("value").toString()),timestamp);
+            updateOdometerObservers( Double.parseDouble(jsonObject.get("value").toString()),timestamp);
         }
     }
